requirements.txt
customtkinter
Pillow
py-cpuinfo
psutil
pywin32
screeninfo